Setup workspace project tests to run in parallel (#788)
